+++ TreeView/TreeModel.cpp
@@ -121,7 +121,7 @@ QModelIndex TreeModel::index(int row, int column, const QModelIndex & parent) co
 {
   if(!parent.isValid())
   {
-    if(row >= m_items.size())
+    if(size_t(row) >= m_items.size())
       return QModelIndex();
 
     QModelIndex idx = m_items[row]->modelIndex();
@@ -134,7 +134,7 @@ QModelIndex TreeModel::index(int row, int column, const QModelIndex & parent) co
   }
 
   TreeItem * parentItem = (TreeItem *)parent.internalPointer();
-  if(row >= parentItem->numChildren())
+  if(size_t(row) >= parentItem->numChildren())
     return QModelIndex();
 
   TreeItem * item = parentItem->child(row);
